refactor: modularizado popover de config (con IA)
